+++ negotiation/src/main/java/it/eng/negotiation/service/policy/PolicyManager.java
@@ -1,4 +1,4 @@
-package it.eng.negotiation.service;
+package it.eng.negotiation.service.policy;
 
 import org.springframework.stereotype.Component;
 
@@ -20,13 +20,13 @@ public class PolicyManager {
 
 	public int getAccessCount(String agreementId) {
 		return repository.findByAgreementId(agreementId).map(pe -> pe.getCount())
-				.orElseThrow(() -> new PolicyEnforcementException("Agreement with id '" + agreementId + "' not found"));
+				.orElseThrow(() -> new PolicyEnforcementException("PolicyManager for agreementId '" + agreementId + "' not found"));
 	}
 	
 	public synchronized void updateAccessCount(String agreementId) {
 		log.info("Updating access count for agreementId {}", agreementId);
 		PolicyEnforcement pe = repository.findByAgreementId(agreementId)
-				.orElseThrow(() -> new PolicyEnforcementException("Agreement with id '" + agreementId + "' not found"));
+				.orElseThrow(() -> new PolicyEnforcementException("PolicyManager for agreementId  '" + agreementId + "' not found"));
 		pe.setCount(pe.getCount() + 1);
 		repository.save(pe);
 		log.debug("Access count updated to {}", pe.getCount());